Make the return data type BigDecimal instead of double

+++ src/main/java/com/dufuna/berlin/imodupsy/tax/service/TaxCalService.java
@@ -1,10 +1,11 @@
 package com.dufuna.berlin.imodupsy.tax.service;
 
+import java.math.BigDecimal;
+
 /**
  * @author iModupsy
  * @created 11/07/2022
  */
 public interface TaxCalService {
-
-    double calculateTax(double income);
+    BigDecimal calculateTax(double income);
 }
